pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mama"
version = "0.1.0"
description = "Marketing Agent Multi-Agent Architecture"
requires-python = ">=3.11"
dependencies = [
    # Agent Framework
    "langgraph>=0.2.0",
    "langchain-anthropic>=0.3.0",
    "langchain-core>=0.3.0",
    "anthropic>=0.45.0",

    # API & Web
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "httpx>=0.28.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",

    # Database
    "sqlalchemy[asyncio]>=2.0.36",
    "alembic>=1.14.0",
    "asyncpg>=0.30.0",
    "psycopg[binary]>=3.2.0",

    # Cache / Queue
    "redis[hiredis]>=5.2.0",
    "celery[redis]>=5.4.0",
    "kombu>=5.4.0",

    # LangGraph persistence
    "langgraph-checkpoint-postgres>=2.0.0",

    # Media processing
    "Pillow>=11.0.0",
    "ffmpeg-python>=0.2.0",
    "moviepy>=1.0.3",

    # Auth
    "bcrypt>=4.0.0",
    "PyJWT>=2.8.0",

    # Multi-LLM fallback
    "openai>=1.0.0",
    "google-genai>=1.0.0",

    # Social Media APIs
    "tweepy>=4.15.0",
    "google-api-python-client>=2.155.0",
    "google-auth>=2.37.0",

    # Utilities
    "python-dotenv>=1.0.1",
    "structlog>=24.4.0",
    "prometheus-client>=0.21.0",
    "boto3>=1.35.0",
    "pytrends>=4.9.2",
    "tenacity>=9.0.0",
    "python-multipart>=0.0.20",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.25.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.9.0",
    "mypy>=1.14.0",
    "pre-commit>=4.1.0",
    "httpx>=0.28.0",
    "factory-boy>=3.3.0",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.hatch.build.targets.wheel]
packages = ["src"]
